Refactoring mantissa and pval creation out of utils and into pval
